2015.07.30. Plate manipulator integration in mxcube
PlateManipulator hwobj is like a sample changer (inherits GenericSampleChanger) and has the same methods and events. It has only load function (no unload), because always plate is in a defined position.
Communication is via exporter channels and commands (MD2/3 jlib software)
Plate size is defined by rows, colls and drops per cell.
In plate-manipulator.xml 8 rows, 12 colls and 3 drops per cell are defined.
Uncomment channels and commands to actual connection to diffractometer
In gui when Plate mode selected then tree is populated according to plate configuration defined in xml.
If in diffractometer plate mode detected (in_plate_mode returns True) then smaple changer should be disabled. Right now if user selects sample changer from the combobox an warning is printed. This should be more generic and combobox should be populated based on diffractometer state.

TODO:
1. Add correct method to detect if diffractometer is in sample changer/plate/minikappa mode. Right now hard coded method in_plate_mode returns if diffractometer is in plate mode. This should be connected to head detection (see EMBLMiniDiff as an example). At beginning this propery could be set via xml
2. Add omega angle limit (snapshots and centring) if plate used. Dynamic limits instead of static limits (see EMBLMiniDiff as an example)
3. Test plates with MD2.
4. Add Crims to get information about plate and crystals. Use crystals as centred positions
5. Develop PlateBrick (in progress)

Files modified:

Bricks/TreeBrick.py: add plate manipulator hwobj and modified logged_in. If method is ok, then logged_in_old_version should be removed
Bricks/widgets/dc_tree_widget.py: added all methods to populate and handle plate in tree view
HardwareObjects/DiffractometerMockup.py: added method in_plate_mode (to indicate if in plate mode)
HardwareObjects/LimaVideo.py: added method change_owner to drop sudo rights
HardwareObjects/MicrodiffMotorMockup.py: inherits AbstractMotor
HardwareObjects/PlateManipulator.py: added mockup style behaviour if no channels and commands defined
HardwareObjects/QueueModel.py: added plate model
HardwareObjects/queue_entry.py: added methods to mount sample in plate mode
HardwareObjects/queue_model_objects_v1.py: modifications to initialize Sample in plate mode

+++ HardwareObjects/PlateManipulator.py
@@ -162,6 +162,7 @@ class PlateManipulator(SampleChanger):
         self.cmd_move_to_location = self.getCommandObject("startMovePlateToLocation")
         self.chan_current_phase = self.getChannelObject("CurrentPhase")
         self.chan_plate_location = self.getChannelObject("PlateLocation")
+        
         self.chan_state = self.getChannelObject("State")
         if self.chan_state is not None:
             self.chan_state.connectSignal("update", self._onStateChanged)
@@ -222,15 +223,33 @@ class PlateManipulator(SampleChanger):
             if (sample!=selected):
                 self._doSelect(sample)
             self._setLoadedSample(sample)
-        #TODO: Add pre-positioning and image matching
 
     def load_sample(self, sample_location=None):
         """
         Descript. : function to move to plate location.
                     Location is estimated by sample location and reference positions.
         """
-        pos_y = float(sample_location[2]) / (self.num_drops + 1)   
-        self.cmd_move_to_location(sample_location[0], sample_location[1] - 1, self.reference_pos_x, pos_y) 
+        row =  sample_location[0]
+        col =  sample_location[1]
+        drop = sample_location[2]
+
+        pos_y = float(drop) / (self.num_drops + 1)   
+        if self.cmd_move_to_location:
+            self.cmd_move_to_location(row, col - 1, self.reference_pos_x, pos_y) 
+        else:
+            #No actual move cmd defined. Act like a mockup
+            cell = self.getComponentByAddress("%s%d" %(chr(65 + row), col))
+            drop = cell.getComponentByAddress("%s%d:%d" %(chr(65 + row), col, drop))
+            new_sample = drop.getSample()
+            old_sample = self.getLoadedSample()
+            new_sample = drop.getSample()
+            if old_sample != new_sample:
+                if old_sample is not None:
+                    old_sample._setLoaded(False, True)
+                if new_sample is not None:
+                    new_sample._setLoaded(True, True)
+
+            #new_sample._setLoaded(True, True)
         
     def _doUnload(self,sample_slot=None):
         """
@@ -282,6 +301,7 @@ class PlateManipulator(SampleChanger):
         Descript. :
         """
         self._updateState()
+        #TODO remove self._updateLoadedSample and add event to self.chan_plate_location
         self._updateLoadedSample()
 
     def _updateState(self):
@@ -303,11 +323,7 @@ class PlateManipulator(SampleChanger):
         """
         plate_location = None
         if self.chan_plate_location is not None:
-            #If plate location event works this could be removed
             plate_location = self.chan_plate_location.getValue()
-
-        #uncomment next lines for testing
-        plate_location = [2, 2, 2, 2]
 
         if plate_location is not None:
             row = int(plate_location[0])
@@ -318,6 +334,9 @@ class PlateManipulator(SampleChanger):
                 drop_index = self.num_drops
 
             cell = self.getComponentByAddress("%s%d" %(chr(65 + row), col + 1))
+
+            if cell is None:
+                return
             old_sample = self.getLoadedSample()
             drop = cell.getComponentByAddress("%s%d:%d" %(chr(65 + row), col + 1, drop_index))
             new_sample = drop.getSample()
